Move existing ownership-related tests to their own file
